fixing compat tests
